refactor: changing directories in types
