+++ api/gateway/admin/src/types/request/index.ts
@@ -1,2 +1,2 @@
-export * from './admin'
-export * from './auth'
+export * from './v1/admin'
+export * from './v1/auth'
