+++ api/gateway/admin/src/types/response/index.ts
@@ -1,4 +1,4 @@
-export * from './admin'
-export * from './auth'
+export * from './v1/admin'
+export * from './v1/auth'
+export * from './v1/user'
 export * from './error'
-export * from './user'
