+++ api/gateway/admin/src/types/response/v1/auth.ts
