+++ api/gateway/admin/src/types/response/v1/user.ts
