+++ api/gateway/native/src/types/request/v1/chat.ts
