+++ api/gateway/native/src/types/response/index.ts
@@ -1,6 +1,6 @@
-export * from './auth'
-export * from './book'
-export * from './bookshelf'
-export * from './chat'
+export * from './v1/auth'
+export * from './v1/book'
+export * from './v1/bookshelf'
+export * from './v1/chat'
+export * from './v1/user'
 export * from './error'
-export * from './user'
