+++ api/gateway/native/src/types/response/v1/book.ts
