+++ api/gateway/native/src/types/response/v1/chat.ts
